+++ presentation/src/main/java/com/lighthouse/presentation/ui/home/HomeFragment.kt
@@ -1,5 +1,6 @@
 package com.lighthouse.presentation.ui.home
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import androidx.annotation.StringRes
@@ -12,9 +13,11 @@ import com.lighthouse.presentation.R
 import com.lighthouse.presentation.databinding.FragmentHomeBinding
 import com.lighthouse.presentation.extension.dp
 import com.lighthouse.presentation.extension.repeatOnStarted
+import com.lighthouse.presentation.extra.Extras
 import com.lighthouse.presentation.ui.common.GifticonViewHolderType
 import com.lighthouse.presentation.ui.common.UiState
 import com.lighthouse.presentation.ui.common.viewBindings
+import com.lighthouse.presentation.ui.detailgifticon.GifticonDetailActivity
 import com.lighthouse.presentation.ui.main.MainViewModel
 import com.lighthouse.presentation.ui.map.adapter.GifticonAdapter
 import com.lighthouse.presentation.util.recycler.ListSpaceItemDecoration
@@ -28,8 +31,20 @@ class HomeFragment : Fragment(R.layout.fragment_home) {
     private val binding: FragmentHomeBinding by viewBindings()
     private val mainViewModel: MainViewModel by activityViewModels()
     private val homeViewModel: HomeViewModel by viewModels()
-    private val nearGifticonAdapter = GifticonAdapter(GifticonViewHolderType.VERTICAL)
-    private val expireGifticonAdapter = GifticonAdapter(GifticonViewHolderType.VERTICAL)
+    private val nearGifticonAdapter = GifticonAdapter(GifticonViewHolderType.VERTICAL) { gifticon ->
+        startActivity(
+            Intent(requireContext(), GifticonDetailActivity::class.java).apply {
+                putExtra(Extras.KEY_GIFTICON_ID, gifticon.id)
+            }
+        )
+    }
+    private val expireGifticonAdapter = GifticonAdapter(GifticonViewHolderType.VERTICAL) { gifticon ->
+        startActivity(
+            Intent(requireContext(), GifticonDetailActivity::class.java).apply {
+                putExtra(Extras.KEY_GIFTICON_ID, gifticon.id)
+            }
+        )
+    }
     private val itemDecoration = ListSpaceItemDecoration(
         space = 8.dp,
         start = 4.dp,
