+++ presentation/src/main/java/com/lighthouse/presentation/ui/map/adapter/GifticonAdapter.kt
@@ -1,14 +1,19 @@
 package com.lighthouse.presentation.ui.map.adapter
 
+import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.lighthouse.domain.model.Gifticon
+import com.lighthouse.presentation.R
 import com.lighthouse.presentation.adapter.BindableListAdapter
+import com.lighthouse.presentation.databinding.ItemGifticonHorizontalBinding
+import com.lighthouse.presentation.databinding.ItemGifticonVerticalBinding
 import com.lighthouse.presentation.ui.common.GifticonViewHolderType
 
 class GifticonAdapter(
-    private val gifticonViewHolderType: GifticonViewHolderType
+    private val gifticonViewHolderType: GifticonViewHolderType,
+    private val onClick: (Gifticon) -> Unit
 ) : BindableListAdapter<Gifticon, RecyclerView.ViewHolder>(diff) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
@@ -22,6 +27,42 @@ class GifticonAdapter(
         when (holder) {
             is GifticonVerticalItemViewHolder -> holder.bind(currentList[position])
             is GifticonHorizontalItemViewHolder -> holder.bind(currentList[position])
+        }
+    }
+
+    inner class GifticonHorizontalItemViewHolder(
+        parent: ViewGroup,
+        private val binding: ItemGifticonHorizontalBinding = ItemGifticonHorizontalBinding.bind(
+            LayoutInflater.from(parent.context).inflate(R.layout.item_gifticon_horizontal, parent, false)
+        )
+    ) : RecyclerView.ViewHolder(binding.root) {
+
+        init {
+            binding.root.setOnClickListener {
+                onClick(currentList[absoluteAdapterPosition])
+            }
+        }
+
+        fun bind(gifticon: Gifticon) {
+            binding.gifticon = gifticon
+        }
+    }
+
+    inner class GifticonVerticalItemViewHolder(
+        parent: ViewGroup,
+        private val binding: ItemGifticonVerticalBinding = ItemGifticonVerticalBinding.bind(
+            LayoutInflater.from(parent.context).inflate(R.layout.item_gifticon_vertical, parent, false)
+        )
+    ) : RecyclerView.ViewHolder(binding.root) {
+
+        init {
+            binding.root.setOnClickListener {
+                onClick(currentList[absoluteAdapterPosition])
+            }
+        }
+
+        fun bind(gifticon: Gifticon) {
+            binding.gifticon = gifticon
         }
     }
